+++ .github/CODEOWNERS
@@ -1,3 +1,4 @@
 * @mure @cameronwaterman
 
-/src/datasources/asset @CiprianAnton @kkerezsi
+/src/datasources/asset-calibration @CiprianAnton @kkerezsi
+/src/datasources/asset-utilization @CiprianAnton @kkerezsi
