+++ src/datasources/asset/module.ts
@@ -1,8 +1,8 @@
 import { DataSourcePlugin } from '@grafana/data';
-import { AssetDataSource } from './AssetDataSource';
-import { AssetQueryEditor } from './components/AssetQueryEditor';
+import { AssetUtilizationDataSource } from './AssetUtilizationDataSource';
+import { AssetUtilizationQueryEditor } from './components/AssetUtilizationQueryEditor';
 import { HttpConfigEditor } from 'core/components/HttpConfigEditor';
 
-export const plugin = new DataSourcePlugin(AssetDataSource)
+export const plugin = new DataSourcePlugin(AssetUtilizationDataSource)
   .setConfigEditor(HttpConfigEditor)
-  .setQueryEditor(AssetQueryEditor);
+  .setQueryEditor(AssetUtilizationQueryEditor);
