+++ src/datasources/asset/plugin.json
@@ -1,6 +1,6 @@
 {
   "type": "datasource",
-  "name": "SystemLink Assets",
+  "name": "SystemLink Assets Utilization",
   "id": "ni-slasset-datasource",
   "metrics": true,
   "info": {
